Added Firebombs which consume wood(fuel)
I've added in the ability to wield firebombs by pressing the 'a' key, which puts you in attack mode and unable to move, and clicking at the target. Currently your range is fairly limited.

Firebombs consume wood as it also counts as your 'fuel' resource. This will be differentiated later.